+++ .editorconfig
@@ -1,5 +1,15 @@
 root = true
 
-[**.js]
-indent_style = tab
-trim_trailing_whitespace=true
+[*]
+end_of_line = lf
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+[*.js]
+charset = utf-8
+indent_style = space
+indent_size = 2
+
+[{package.json,.babelrc,.eslintrc}]
+indent_style = space
+indent_size = 2
