+++ deprecated/example/entry.js
