+++ package.json
@@ -27,12 +27,12 @@
   "scripts": {
     "check": "npm run lint && npm test",
     "lint": "eslint .",
-    "test": "jasmine"
+    "test": "ava tests/"
   },
   "devDependencies": {
+    "ava": "^0.15.2",
     "eslint": "^3.1.1",
     "grunt": "^1.0.0",
-    "jasmine": "2.3.2",
     "webpack": "^1.13.0",
     "webpack-dev-server": "^1.14.1"
   }
